Failed splitting, cannot reach file_io from umi_external
